+++ api/src/main/kotlin/handler/EvHandler.kt
@@ -1,6 +1,6 @@
 package com.wafflestudio.snutt.handler
 
-import com.wafflestudio.snutt.lectures.service.LectureService
+import com.wafflestudio.snutt.evaluation.service.EvService
 import com.wafflestudio.snutt.middleware.SnuttRestApiNoAuthMiddleware
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.server.ServerRequest
@@ -8,7 +8,7 @@ import org.springframework.web.reactive.function.server.ServerResponse
 
 @Component
 class EvHandler(
-    private val lectureService: LectureService,
+    private val evService: EvService,
     snuttRestApiNoAuthMiddleware: SnuttRestApiNoAuthMiddleware,
 ) : ServiceHandler(
         handlerMiddleware = snuttRestApiNoAuthMiddleware,
@@ -16,6 +16,6 @@ class EvHandler(
     suspend fun getLectureEvaluationSummary(req: ServerRequest): ServerResponse =
         handle(req) {
             val lectureId = req.pathVariable("lectureId")
-            lectureService.getEvSummary(lectureId)
+            evService.getEvSummary(lectureId)
         }
 }
